requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
Faker
django-paranoid
requests
django-debug-toolbar
djangorestframework
flake8
django-rest-auth[with_social]
django-rest-swagger

